strategies/stock-screener/requirements.txt
yfinance>=0.2.44
openbb>=4.0.0
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0

# Optional news providers
finnhub-python>=2.4.20

# Dev
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Trading bot dependencies
alpaca-py>=0.43.0
APScheduler>=3.10.0
